+++ client/src/app/constants/app_tabs.js
@@ -5,3 +5,4 @@ export const SETTINGS = 'settings';
 export const INOCULATIONS = 'inoculations';
 export const INFECTIONS = 'infections';
 export const UPLOAD = 'upload';
+export const APPLICATION_SETTINGS = 'application_settings';