+++ client/src/app/containers/AppNavbar.js
@@ -20,7 +20,8 @@ import {
     WEIGHT,
     INOCULATIONS,
     INFECTIONS,
-    UPLOAD
+    UPLOAD,
+    APPLICATION_SETTINGS
 } from '../constants/app_tabs';
 import {
     MdInfo
@@ -104,6 +105,14 @@ export class AppNavbar extends React.Component {
 
         changeActiveTab(UPLOAD);
     }
+    @autobind
+    onApplicationSettingsClick() {
+        const {
+            changeActiveTab
+        } = this.props;
+
+        changeActiveTab(APPLICATION_SETTINGS);
+    }
     render() {
         const {
             username,
@@ -142,8 +151,9 @@ export class AppNavbar extends React.Component {
                 </Nav>
                 <Nav className="navbar-settings" pullRight>
                     <NavDropdown eventKey={4} title="Settings" id="settings-dropdown" disabled={isFetchingData}>
-                        <MenuItem eventKey={4.1} onClick={this.handleSettingsChangeClick}>Change info</MenuItem>
+                        <MenuItem eventKey={4.1} onClick={this.handleSettingsChangeClick}>Child info</MenuItem>
                         <MenuItem eventKey={4.2} onClick={this.onUploadTabClick}>Upload image</MenuItem>
+                        <MenuItem eventKey={4.3} onClick={this.onApplicationSettingsClick}>App settings</MenuItem>
                         <MenuItem divider/>
                         <MenuItem eventKey={4.2} onClick={logout}>Logout</MenuItem>
                     </NavDropdown>